Add ASG example
Example showing the case when PV allowing write comes from the another IOC.
In this case the copy PV reading from the first IOC is used.
If the original access PV fails (e.g. IOC is stopped),
the 0 value is forced so every user is allowed to write to controlled PV.